requirements.txt
# Core dependencies
python-dotenv>=1.0.0
httpx>=0.25.0

# Notion API
notion-client>=2.2.0

# AI/ML
anthropic>=0.40.0
openai>=1.0.0

# Audio transcription
openai-whisper>=20231117
# OR for API-only: whisper not needed, uses openai client

# Embeddings & similarity
numpy>=1.24.0
scikit-learn>=1.3.0

# File watching
watchdog>=3.0.0

# CLI
click>=8.1.0
rich>=13.0.0

# Async support
asyncio-throttle>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Local whisper (large download)
# torch>=2.0.0
# torchaudio>=2.0.0
